Split filesystem module into separate ones
