+++ src/filesystem/cluster.rs
@@ -1,0 +1,44 @@
+/// Represents a cluster on disk.
+#[cfg_attr(feature = "defmt-log", derive(defmt::Format))]
+#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+pub struct Cluster(pub(crate) u32);
+
+impl Cluster {
+    /// Magic value indicating an invalid cluster value.
+    pub const INVALID: Cluster = Cluster(0xFFFF_FFF6);
+    /// Magic value indicating a bad cluster.
+    pub const BAD: Cluster = Cluster(0xFFFF_FFF7);
+    /// Magic value indicating a empty cluster.
+    pub const EMPTY: Cluster = Cluster(0x0000_0000);
+    /// Magic value indicating the cluster holding the root directory (which
+    /// doesn't have a number in FAT16 as there's a reserved region).
+    pub const ROOT_DIR: Cluster = Cluster(0xFFFF_FFFC);
+    /// Magic value indicating that the cluster is allocated and is the final cluster for the file
+    pub const END_OF_FILE: Cluster = Cluster(0xFFFF_FFFF);
+}
+
+impl core::ops::Add<u32> for Cluster {
+    type Output = Cluster;
+    fn add(self, rhs: u32) -> Cluster {
+        Cluster(self.0 + rhs)
+    }
+}
+
+impl core::ops::AddAssign<u32> for Cluster {
+    fn add_assign(&mut self, rhs: u32) {
+        self.0 += rhs;
+    }
+}
+
+impl core::ops::Add<Cluster> for Cluster {
+    type Output = Cluster;
+    fn add(self, rhs: Cluster) -> Cluster {
+        Cluster(self.0 + rhs.0)
+    }
+}
+
+impl core::ops::AddAssign<Cluster> for Cluster {
+    fn add_assign(&mut self, rhs: Cluster) {
+        self.0 += rhs.0;
+    }
+}
